requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
playwright==1.41.2
python-multipart==0.0.9
httpx==0.26.0
dnspython==2.5.0
Pillow==10.2.0

